+++ packages/showcases/react/stories/components/selection-controls/VtmnChip/VtmnChip.stories.tsx
@@ -10,7 +10,17 @@ import {
 export default {
   title: 'Components / Selection controls / VtmnChip',
   component: VtmnChip,
-  argTypes,
+  argTypes: {
+    ...argTypes,
+    children: {
+      type: { name: 'string', required: false },
+      description: 'The content to render inside the component.',
+      defaultValue: 'Button',
+      control: {
+        type: 'text',
+      },
+    },
+  },
   parameters,
 } as Meta;