+++ packages/showcases/svelte/stories/components/indicators/VtmnRating/VtmnRating.stories.svelte
@@ -45,20 +45,20 @@
 </Story>
 
 <Story name="Read Only with text">
-  <VtmnRating class="rating-stories" readOnly compact rating={1}>
+  <VtmnRating class="rating-stories" readonly compact rating={1}>
     <span slot="primary">4.1/5</span>
   </VtmnRating>
-  <VtmnRating class="rating-stories" readOnly compact rating={0.1}>
+  <VtmnRating class="rating-stories" readonly compact rating={0.1}>
     <span slot="primary">1/5</span>
   </VtmnRating>
-  <VtmnRating class="rating-stories" size="small" readOnly compact rating={1}>
+  <VtmnRating class="rating-stories" size="small" readonly compact rating={1}>
     <span slot="primary">2.9/5</span>
     <span slot="secondary">(74)</span>
   </VtmnRating>
-  <VtmnRating class="rating-stories" emphasis readOnly compact rating={1}>
+  <VtmnRating class="rating-stories" emphasis readonly compact rating={1}>
     <span slot="primary">74 ratings</span>
   </VtmnRating>
-  <VtmnRating class="rating-stories" emphasis readOnly rating={2.3}>
+  <VtmnRating class="rating-stories" emphasis readonly rating={2.3}>
     <span slot="primary">2.3/5</span>
     <span slot="secondary">(74)</span>
   </VtmnRating>
@@ -67,19 +67,19 @@
 <Story name="Read only">
   <div class="rating-story-align">
     <div>
-      <VtmnRating class="rating-stories" readOnly size="small" rating={0} />
-      <VtmnRating class="rating-stories" readOnly size="small" rating={0.5} />
-      <VtmnRating class="rating-stories" readOnly size="small" rating={1} />
-      <VtmnRating class="rating-stories" readOnly size="small" rating={1.5} />
-      <VtmnRating class="rating-stories" readOnly size="small" rating={2} />
-      <VtmnRating class="rating-stories" readOnly size="small" rating={2.5} />
+      <VtmnRating class="rating-stories" readonly size="small" rating={0} />
+      <VtmnRating class="rating-stories" readonly size="small" rating={0.5} />
+      <VtmnRating class="rating-stories" readonly size="small" rating={1} />
+      <VtmnRating class="rating-stories" readonly size="small" rating={1.5} />
+      <VtmnRating class="rating-stories" readonly size="small" rating={2} />
+      <VtmnRating class="rating-stories" readonly size="small" rating={2.5} />
     </div>
     <div>
-      <VtmnRating class="rating-stories" readOnly emphasis rating={3} />
-      <VtmnRating class="rating-stories" readOnly emphasis rating={3.5} />
-      <VtmnRating class="rating-stories" readOnly emphasis rating={4} />
-      <VtmnRating class="rating-stories" readOnly emphasis rating={4.5} />
-      <VtmnRating class="rating-stories" readOnly emphasis rating={5} />
+      <VtmnRating class="rating-stories" readonly emphasis rating={3} />
+      <VtmnRating class="rating-stories" readonly emphasis rating={3.5} />
+      <VtmnRating class="rating-stories" readonly emphasis rating={4} />
+      <VtmnRating class="rating-stories" readonly emphasis rating={4.5} />
+      <VtmnRating class="rating-stories" readonly emphasis rating={5} />
     </div>
   </div>
 </Story>