+++ packages/sources/css/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.62.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.62.1...@vtmn/css@0.62.2) (2022-03-24)
+
+
+### Bug Fixes
+
+* **@vtmn/css:** modal indicator overflow ([#1068](https://github.com/Decathlon/vitamin-web/issues/1068)) ([6338cd2](https://github.com/Decathlon/vitamin-web/commit/6338cd2c0450ecdb32e7d7fb2b6ae2d81847fa4b))
+
+
+
+
+
+## [0.62.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.62.0...@vtmn/css@0.62.1) (2022-03-23)
+
+
+### Bug Fixes
+
+* **@vtmn/css:** `toggle`, incorrect checked icon color ([#1066](https://github.com/Decathlon/vitamin-web/issues/1066)) ([34df2a8](https://github.com/Decathlon/vitamin-web/commit/34df2a833a38d19b22a2c5ad6990f9f94af9d56a))
+
+
+
+
+
 # [0.62.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.61.0...@vtmn/css@0.62.0) (2022-03-22)