+++ packages/sources/svelte/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.28.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.27.0...@vtmn/svelte@0.28.0) (2022-03-24)
+
+
+### Features
+
+* **@vtmn/svelte:** add `VtmnQuantity` component ([#1055](https://github.com/Decathlon/vitamin-web/issues/1055)) ([134fa8a](https://github.com/Decathlon/vitamin-web/commit/134fa8a79f7a0edae885a3ea391be6f282ff987b))
+
+
+
+
+
+# [0.27.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.26.0...@vtmn/svelte@0.27.0) (2022-03-24)
+
+
+### Features
+
+* **@vtmn/svelte:** add `VtmnNavbar` component ([#1044](https://github.com/Decathlon/vitamin-web/issues/1044)) ([f042102](https://github.com/Decathlon/vitamin-web/commit/f042102d61bb8c7a8efc61c5bbc0cf6506d75734))
+
+
+
+
+
+# [0.26.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.25.2...@vtmn/svelte@0.26.0) (2022-03-24)
+
+
+### Features
+
+* **@vtmn/svelte:** add `VtmnTooltip` component ([#1062](https://github.com/Decathlon/vitamin-web/issues/1062)) ([fa81c38](https://github.com/Decathlon/vitamin-web/commit/fa81c387609c352ec18f317ca1afd7696eb4aeb5))
+
+
+
+
+
+## [0.25.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.25.1...@vtmn/svelte@0.25.2) (2022-03-24)
+
+**Note:** Version bump only for package @vtmn/svelte
+
+
+
+
+
+## [0.25.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.25.0...@vtmn/svelte@0.25.1) (2022-03-23)
+
+**Note:** Version bump only for package @vtmn/svelte
+
+
+
+
+
 # [0.25.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.24.2...@vtmn/svelte@0.25.0) (2022-03-23)